requirements.txt
streamlit==1.29.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3
plotly==5.17.0
fpdf2==2.7.6
json5==0.9.14
pydantic==2.5.2
typing-extensions==4.9.0